# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "UrJersey API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/shop/catalogue": {
   "get": {
    "summary": "Catalogue",
    "operationId": "catalogue_api_v1_shop_catalogue_get",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/shop/quote": {
   "post": {
    "summary": "Shop Quote",
    "operationId": "shop_quote_api_v1_shop_quote_post",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/QuoteRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/shop/enquiries": {
   "post": {
    "summary": "Enquiry",
    "description": "Bulk or custom request from the website: becomes a customer, a lead and a task for sales.",
    "operationId": "enquiry_api_v1_shop_enquiries_post",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EnquiryIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/otp/request": {
   "post": {
    "summary": "Otp Request",
    "operationId": "otp_request_api_v1_auth_otp_request_post",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OtpRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/otp/verify": {
   "post": {
    "summary": "Otp Verify",
    "operationId": "otp_verify_api_v1_auth_otp_verify_post",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OtpVerify"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/logout": {
   "post": {
    "summary": "Logout",
    "operationId": "logout_api_v1_auth_logout_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/me": {
   "get": {
    "summary": "Me",
    "operationId": "me_api_v1_me_get",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "summary": "Update Me",
    "operationId": "update_me_api_v1_me_patch",
    "parameters": [
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProfileIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/me/orders": {
   "get": {
    "summary": "My Orders",
    "operationId": "my_orders_api_v1_me_orders_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/me/orders/{order_id}": {
   "get": {
    "summary": "My Order",
    "operationId": "my_order_api_v1_me_orders__order_id__get",
    "parameters": [
     {
      "name": "order_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Order Id"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/me/orders/{order_id}/reorder": {
   "post": {
    "summary": "Reorder",
    "operationId": "reorder_api_v1_me_orders__order_id__reorder_post",
    "parameters": [
     {
      "name": "order_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Order Id"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/orders/{order_id}/invoice": {
   "get": {
    "summary": "Invoice",
    "operationId": "invoice_api_v1_orders__order_id__invoice_get",
    "parameters": [
     {
      "name": "order_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Order Id"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "text/html": {
        "schema": {
         "type": "string"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/orders/{order_id}/track": {
   "get": {
    "summary": "Track",
    "description": "Guest tracking: order id plus the phone number used for it.",
    "operationId": "track_api_v1_orders__order_id__track_get",
    "parameters": [
     {
      "name": "order_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Order Id"
      }
     },
     {
      "name": "phone",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 6,
       "maxLength": 24,
       "title": "Phone"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/me/designs/{did}": {
   "delete": {
    "summary": "Delete Design",
    "operationId": "delete_design_api_v1_me_designs__did__delete",
    "parameters": [
     {
      "name": "did",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Did"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "OtpRequest": {
    "properties": {
     "phone": {
      "type": "string",
      "maxLength": 24,
      "minLength": 6,
      "title": "Phone"
     }
    },
    "type": "object",
    "required": [
     "phone"
    ],
    "title": "OtpRequest"
   },
   "OtpVerify": {
    "properties": {
     "phone": {
      "type": "string",
      "maxLength": 24,
      "minLength": 6,
      "title": "Phone"
     },
     "code": {
      "type": "string",
      "maxLength": 8,
      "minLength": 4,
      "title": "Code"
     },
     "name": {
      "type": "string",
      "maxLength": 80,
      "title": "Name",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "phone",
     "code"
    ],
    "title": "OtpVerify"
   },
...
   "QuoteRequest": {
    "properties": {
     "garment": {
      "type": "string",
      "enum": [
       "jersey",
       "vneck",
       "shorts"
      ],
      "title": "Garment",
      "default": "jersey"
     },
     "fabric": {
      "type": "string",
      "title": "Fabric",
      "default": "standard"
     },
     "logos": {
      "type": "integer",
      "maximum": 4.0,
      "minimum": 0.0,
      "title": "Logos",
      "default": 0
     },
     "lines": {
      "items": {
       "$ref": "#/components/schemas/PriceLine"
      },
      "type": "array",
      "maxItems": 500,
      "minItems": 1,
      "title": "Lines"
     },
     "delivery": {
      "$ref": "#/components/schemas/DeliveryChoice"
     },
     "rush": {
      "type": "boolean",
      "title": "Rush",
      "default": false
     },
     "coupon": {
      "type": "string",
      "maxLength": 24,
      "title": "Coupon",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "lines"
    ],
    "title": "QuoteRequest"
   },
...
  }
 }
}